requirements.txt
gradio>=4,<5
torch
torchvision
numpy
Pillow
huggingface_hub>=0.20
